Move grid-wide nudge repair passes off the grid class
Turn apply_positive_z_plan_to_existing_cells and the two repair passes that
follow it into free functions in nudge_apply, with thin delegators left on grid
so TouchTerrainEarthEngine and the tests are unchanged.

Rewrite self as target throughout, including the bare self inside
getattr(self, "tile_info", None). That lookup is deliberately tolerant because
tests build a grid carrying only cells, and an attribute-only rewrite would
have left a dangling name on the threaded worker-count path.

Emitted meshes are unchanged: all 66 meshes across all 48 launch configs stay
byte identical.
